backend-ai/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic-settings
openai
pandas 
scikit-learn